+++ tests/test_feedstock_io.py
@@ -26,10 +26,7 @@ def parameterize():
         lambda pth, tmp_dir: os.path.relpath(pth, tmp_dir),
         lambda pth, tmp_dir: pth,
     ]:
-        for get_repo in [
-            lambda tmp_dir: None,
-            lambda tmp_dir: git.Repo.init(tmp_dir),
-        ]:
+        for get_repo in [lambda tmp_dir: None, lambda tmp_dir: git.Repo.init(tmp_dir)]:
             try:
                 tmp_dir = tempfile.mkdtemp()
                 keep_dir(tmp_dir)
@@ -37,11 +34,7 @@ def parameterize():
                 old_dir = os.getcwd()
                 os.chdir(tmp_dir)
 
-                yield (
-                    tmp_dir,
-                    get_repo(tmp_dir),
-                    lambda pth: pathfunc(pth, tmp_dir),
-                )
+                yield (tmp_dir, get_repo(tmp_dir), lambda pth: pathfunc(pth, tmp_dir))
             finally:
                 os.chdir(old_dir)
                 shutil.rmtree(tmp_dir)
@@ -64,9 +57,7 @@ class TestFeedstockIO(unittest.TestCase):
             if repo is None:
                 self.assertTrue(fio.get_repo(pathfunc(tmp_dir)) is None)
             else:
-                self.assertIsInstance(
-                    fio.get_repo(pathfunc(tmp_dir)), git.Repo
-                )
+                self.assertIsInstance(fio.get_repo(pathfunc(tmp_dir)), git.Repo)
 
     def test_set_exe_file(self):
         perms = [stat.S_IXUSR, stat.S_IXGRP, stat.S_IXOTH]
@@ -77,9 +68,7 @@ class TestFeedstockIO(unittest.TestCase):
             for tmp_dir, repo, pathfunc in parameterize():
                 filename = "test.txt"
                 filename = os.path.join(tmp_dir, filename)
-                with io.open(
-                    filename, "w", encoding="utf-8", newline="\n"
-                ) as fh:
+                with io.open(filename, "w", encoding="utf-8", newline="\n") as fh:
                     fh.write("")
                 if repo is not None:
                     repo.index.add([filename])
@@ -90,9 +79,7 @@ class TestFeedstockIO(unittest.TestCase):
                 self.assertEqual(file_mode & set_mode, int(set_exe) * set_mode)
                 if repo is not None:
                     blob = next(repo.index.iter_blobs(BlobFilter(filename)))[1]
-                    self.assertEqual(
-                        blob.mode & set_mode, int(set_exe) * set_mode
-                    )
+                    self.assertEqual(blob.mode & set_mode, int(set_exe) * set_mode)
 
     def test_write_file(self):
         for tmp_dir, repo, pathfunc in parameterize():
@@ -146,9 +133,7 @@ class TestFeedstockIO(unittest.TestCase):
 
                 filename = os.path.join(tmp_dir, filename)
 
-                with io.open(
-                    filename, "w", encoding="utf-8", newline="\n"
-                ) as fh:
+                with io.open(filename, "w", encoding="utf-8", newline="\n") as fh:
                     fh.write("")
                 if repo is not None:
                     repo.index.add([filename])
@@ -158,9 +143,7 @@ class TestFeedstockIO(unittest.TestCase):
                     self.assertTrue(os.path.exists(dirname))
                     self.assertTrue(os.path.exists(os.path.dirname(dirname)))
                 if repo is not None:
-                    self.assertTrue(
-                        list(repo.index.iter_blobs(BlobFilter(filename)))
-                    )
+                    self.assertTrue(list(repo.index.iter_blobs(BlobFilter(filename))))
 
                 fio.remove_file(pathfunc(filename))
 
@@ -169,9 +152,7 @@ class TestFeedstockIO(unittest.TestCase):
                     self.assertFalse(os.path.exists(dirname))
                     self.assertFalse(os.path.exists(os.path.dirname(dirname)))
                 if repo is not None:
-                    self.assertFalse(
-                        list(repo.index.iter_blobs(BlobFilter(filename)))
-                    )
+                    self.assertFalse(list(repo.index.iter_blobs(BlobFilter(filename))))
 
     def test_copy_file(self):
         for tmp_dir, repo, pathfunc in parameterize():
@@ -188,18 +169,14 @@ class TestFeedstockIO(unittest.TestCase):
             self.assertTrue(os.path.exists(filename1))
             self.assertFalse(os.path.exists(filename2))
             if repo is not None:
-                self.assertFalse(
-                    list(repo.index.iter_blobs(BlobFilter(filename2)))
-                )
+                self.assertFalse(list(repo.index.iter_blobs(BlobFilter(filename2))))
 
             fio.copy_file(pathfunc(filename1), pathfunc(filename2))
 
             self.assertTrue(os.path.exists(filename1))
             self.assertTrue(os.path.exists(filename2))
             if repo is not None:
-                self.assertTrue(
-                    list(repo.index.iter_blobs(BlobFilter(filename2)))
-                )
+                self.assertTrue(list(repo.index.iter_blobs(BlobFilter(filename2))))
 
             read_text = ""
             with io.open(filename2, "r", encoding="utf-8") as fh: